fix game over bug
