+++ src/gameManager/PlayerModel.js
@@ -2,7 +2,7 @@ import { v4 as uuidv4 } from 'uuid';
 
 export default class PlayerModel {
     constructor(spawnLocations) {
-      this.health = 10;
+      this.health = 2;
       this.maxHealth = 10;
       this.gold = 0;
       this.id = `player-${uuidv4()}`;
@@ -21,9 +21,9 @@ export default class PlayerModel {
       if (this.health > 10) this.health = 10;
     }
   
-    respawn() {
-      this.health = this.maxHealth;
-      const location = this.spawnLocations[Math.floor(Math.random() * this.spawnLocations.length)];
-      [this.x, this.y] = location;
-    }
+    // respawn() {
+    //   this.health = this.maxHealth;
+    //   const location = this.spawnLocations[Math.floor(Math.random() * this.spawnLocations.length)];
+    //   [this.x, this.y] = location;
+    // }
   }
