+++ src/gameManager/utils.js
@@ -6,3 +6,5 @@ export const SpawnerType = {
   export const randomNumber = (min, max) => {
     return Math.floor(Math.random() * max) + min;
   }
+
+
