+++ src/index.js
@@ -7,8 +7,10 @@ import TitleScene from './scenes/TitleScene';
 import OptionsScene from './scenes/OptionsScene';
 import CreditsScene from './scenes/CreditsScene';
 import InstructionScene from './scenes/InstructionScene';
+import LeaderBoardScene from './scenes/LeaderBoardScene';
 import UiScene from './scenes/UiScene';
 import Model from './classes/Model';
+import GameOverScene from './scenes/GameOverScene';
 
  
 
@@ -23,6 +25,8 @@ class Game extends Phaser.Game {
     this.scene.add('Options', OptionsScene);
     this.scene.add('Credits', CreditsScene);
     this.scene.add('Instructions', InstructionScene);
+    this.scene.add('GameOver', GameOverScene);
+    this.scene.add('Score', LeaderBoardScene);
     this.scene.add('Game', GameScene);
     this.scene.add('Ui', UiScene);
     this.scene.start('Boot');
