+++ src/scenes/InstructionScene.js
@@ -29,32 +29,9 @@ this.text = this.add.text(300, 40, 'How to play😎', { fontSize: 40 });
     this.downButton = new Button(this, 300, 300, 'blueButton1', 'blueButton2', '⬇️');
     this.downText = this.add.text(400, 280, 'Move the player down', { fontSize: 24 });
 
-    this.playerName = this.add.text(215, 400, 'Enter your name: ', { fontSize: 20, fontFamily: 'monospace' });
-
-    const input = this.add.dom(480, 412, 'input', {
-      type: 'text',
-      name: 'nameField',
-      fontSize: '32px',
-      backgroundColor: '#fff',
-    });
-    input.scaleX = 0.4;
-    input.scaleY = 0.6;
-
-    const style = 'background: url(assets/ui/blue_button02.png); border: none; border-radius: 3px; color: #fff;';
-    const gamePlay = this.add.dom(590, 412, 'button', style, 'Play');
-    gamePlay.scaleX = 1.5;
-    gamePlay.scaleY = 1.7;
-    gamePlay.addListener('click');
-
-    gamePlay.on('click', () => {
-      if (input.node.value) {
-        this.model = this.sys.game.globals.model;
-        this.model.userName = input.node.value;
-        this.scene.start('Game');
-      }
-    });
-    
+    this.startButton = new Button(this, 600, 500, 'blueButton1', 'blueButton2', 'Start', 'Game')
     
     this.menuButton = new Button(this, 400, 500, 'blueButton1', 'blueButton2', 'Menu', 'Title');
+
 }
 }
